auto_db_github_repo/auto_db_github_repo/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
mysql-connector-python==9.0.0
beautifulsoup4==4.12.3
lxml==5.2.2
